# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Personal Finance Advisor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  income: "income",
  expense: "expense",
} as const;

export interface Transaction {
  id: number;
  amount: number;
  type: TransactionType;
  category: string;
  description: string;
  date: string;
  createdAt: string;
}

export type TransactionInputType =
  (typeof TransactionInputType)[keyof typeof TransactionInputType];

export const TransactionInputType = {
  income: "income",
  expense: "expense",
} as const;

export interface TransactionInput {
  amount: number;
  type: TransactionInputType;
  category: string;
  description: string;
  date: string;
}

export type TransactionUpdateType =
  (typeof TransactionUpdateType)[keyof typeof TransactionUpdateType];

export const TransactionUpdateType = {
  income: "income",
  expense: "expense",
} as const;

export interface TransactionUpdate {
  amount?: number;
  type?: TransactionUpdateType;
  category?: string;
  description?: string;
  date?: string;
}

export type BudgetPeriod = (typeof BudgetPeriod)[keyof typeof BudgetPeriod];

export const BudgetPeriod = {
  monthly: "monthly",
  weekly: "weekly",
} as const;

export interface Budget {
  id: number;
  category: string;
  limit: number;
  period: BudgetPeriod;
  createdAt: string;
}

export type BudgetInputPeriod =
  (typeof BudgetInputPeriod)[keyof typeof BudgetInputPeriod];

export const BudgetInputPeriod = {
  monthly: "monthly",
  weekly: "weekly",
} as const;

export interface BudgetInput {
  category: string;
  limit: number;
  period: BudgetInputPeriod;
}

export type BudgetUpdatePeriod =
  (typeof BudgetUpdatePeriod)[keyof typeof BudgetUpdatePeriod];

export const BudgetUpdatePeriod = {
  monthly: "monthly",
  weekly: "weekly",
} as const;

export interface BudgetUpdate {
  category?: string;
  limit?: number;
  period?: BudgetUpdatePeriod;
}

export interface Category {
  id: number;
  name: string;
  icon: string;
  color: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  id: number;
  role: ChatMessageRole;
  content: string;
  createdAt: string;
}

export interface ChatInput {
  content: string;
}

export interface FinancialSummary {
  totalIncome: number;
  totalExpenses: number;
  netSavings: number;
  savingsRate: number;
  transactionCount: number;
  thisMonthExpenses: number;
  thisMonthIncome: number;
}

export interface CategorySpending {
  category: string;
  total: number;
  percentage: number;
  color: string;
}

export interface MonthlyTrend {
  month: string;
  income: number;
  expenses: number;
  savings: number;
}

export type SpendingPredictionsRiskLevel =
  (typeof SpendingPredictionsRiskLevel)[keyof typeof SpendingPredictionsRiskLevel];

export const SpendingPredictionsRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface SpendingPredictions {
  predictedMonthlyExpense: number;
  confidenceLevel: number;
  riskLevel: SpendingPredictionsRiskLevel;
  alerts: string[];
  recommendations: string[];
}

export type BudgetStatusStatus =
  (typeof BudgetStatusStatus)[keyof typeof BudgetStatusStatus];

export const BudgetStatusStatus = {
  safe: "safe",
  warning: "warning",
  exceeded: "exceeded",
} as const;

export interface BudgetStatus {
  id: number;
  category: string;
  limit: number;
  spent: number;
  remaining: number;
  percentage: number;
  period: string;
  status: BudgetStatusStatus;
}

export type ListTransactionsParams = {
  category?: string;
  type?: ListTransactionsType;
  limit?: number;
};

export type ListTransactionsType =
  (typeof ListTransactionsType)[keyof typeof ListTransactionsType];

export const ListTransactionsType = {
  income: "income",
  expense: "expense",
} as const;