# lib/api-zod/src/generated/types/budgetInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Personal Finance Advisor API
 * OpenAPI spec version: 0.1.0
 */
import type { BudgetInputPeriod } from "./budgetInputPeriod";

export interface BudgetInput {
  category: string;
  limit: number;
  period: BudgetInputPeriod;
}