# lib/api-zod/src/generated/types/budgetStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Personal Finance Advisor API
 * OpenAPI spec version: 0.1.0
 */
import type { BudgetStatusStatus } from "./budgetStatusStatus";

export interface BudgetStatus {
  id: number;
  category: string;
  limit: number;
  spent: number;
  remaining: number;
  percentage: number;
  period: string;
  status: BudgetStatusStatus;
}