# lib/api-zod/src/generated/types/budgetUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Personal Finance Advisor API
 * OpenAPI spec version: 0.1.0
 */
import type { BudgetUpdatePeriod } from "./budgetUpdatePeriod";

export interface BudgetUpdate {
  category?: string;
  limit?: number;
  period?: BudgetUpdatePeriod;
}