# lib/api-zod/src/generated/types/spendingPredictions.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Personal Finance Advisor API
 * OpenAPI spec version: 0.1.0
 */
import type { SpendingPredictionsRiskLevel } from "./spendingPredictionsRiskLevel";

export interface SpendingPredictions {
  predictedMonthlyExpense: number;
  confidenceLevel: number;
  riskLevel: SpendingPredictionsRiskLevel;
  alerts: string[];
  recommendations: string[];
}