# lib/api-zod/src/generated/types/transactionUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Personal Finance Advisor API
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionUpdateType } from "./transactionUpdateType";

export interface TransactionUpdate {
  amount?: number;
  type?: TransactionUpdateType;
  category?: string;
  description?: string;
  date?: string;
}